requirements_langgraph.txt
# LangGraph Workflow Requirements
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Existing requirements
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fpdf>=1.7.2
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
spacy>=3.7.0

# For graph visualization
graphviz>=0.20.0
matplotlib>=3.7.0

# For testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
